Increase the max number of supported vCPUs on Intel KVM hosts.
At the moment, Cloud Hypervisor uses u8 for CPU IDs and for
topology components such as threads_per_core and cores_per_die.
This essentially limits the max number of vCPUs to 255 on all
supported targets/architectures. In addition, mptable
supports at most 254 CPUs on x86_64.

This patch raises this limit on Intel KVM x86_64 hosts;
the focus on Intel KVM (Linux) hosts is driven by practical
considerations (availability of suitable testing hardware)
than any theoretical/technical limitations.

More specifically:

- (v)CPU IDs/numbers are now u32 instead of u8;
- topology components are now u16 instead of u8;
  - cpuid on Intel dedicates 16 bits to topology components;
  - allows having a flat (1, N, 1, 1) topology for N > 255;
- mptable is NOT created on x86_64 if the max number of
  (v)CPUs exceeds 254, without failing setup_mptable();
- topology is now passed around as a 4-tuple instead of
  a 3-tuple (all four topology components can be larger
  than one on modern hardware and large VMs that emulate
  or virtualize it);
- arch/src/x86_64/mod.rs:
  - x2 APIC is enabled;
  - vCPUs are booted in x2 APIC mode if max APIC ID > 254;
  - KVM_FEATURE_MSI_EXT_DEST_ID is enabled, with
    corresponding destination translations;
  - CPUID topology bits are tweaked a bit;
- vmm/src/config.rs and a couple of other places make sure
  that the limit is raised only on Intel KVM hosts.

The changes above appear to be the minimum required to let
a modern Linux Kernel successfully boot with more than 256
CPUs on Intel KVM hosts (either in a flat or a
hierarchical/NUMA mode).

Tested on an Intel host with 480 logical CPUs:
- a flat 1:320:1:1 topology works;
- a 2-level 2:80:1:4 topology works;
- a 3-level 2:40:2:4 topology works.

It is worth calling out that BusDevice trait, implemented
by CpuManager in vmm/src/cpu.rs, uses u8 for CPU IDs, and
this is not changing in this patchset, as the trait
serves as an ABI for CPU hotplugging.

CPU hotplugging support for CPUs with CPU IDs larger than
255 is out of scope here.

Changes introduced here should probably be considered
experimental (subject to further testing).

Signed-off-by: Barret Rhoden <brho@google.com>
Signed-off-by: Neel Natu <neelnatu@google.com>
Signed-off-by: Ofir Weisse <oweisse@google.com>
Signed-off-by: Peter Oskolkov <posk@google.com>